udev (class) support for 2.6.26 - 2.6.30, breaking prior support
The old class/udev support was disabled anyway, as it
wasn't working on kernels >= 2.6.26. I replaced it
completely with a more modular design, that should make it
possible to re-add support for < 2.6.26 if we need it
again. I compile-tested it on 2.6.26 and 2.6.30.

WARNING: This was only tested on a desktop system without
the actual hardware. The test on real hardware is still
pending. (And this commit makes this testing easier.)