Pulseblaster big change: Experimental support for SP17 board
The old design was geared towards the AMCC way of accessing
the boards. Even the debug board was an amcc board with
pciboard==NULL / baseaddr==0.

Instead, there is now a "boardtype" field in the device and
an appropiate enum. It can have three values: PBB_DEBUG,
PBB_GENERIC_AMCC, and PBB_GENERIC_PCI. For reading and
writing registers, there are now dispatchers which take the
device to work on.

I have reinvented the baseaddr in the device, currently as
somewhat like a "cache". When we pass around a
my_dev-pointer anyway, it's easier that way.

This whole change compiles. Testing on real hardware is the
next step.