Quick start at Eurotherm driver for python
This is a quick hack by Marcus to communicate with the
Eurotherm from python. Everything is hardcoded, so don't
expect to be able to use this directly.

This is here, so that it's not lost, and so we can work on
it and make a nicer interface. So don't expect the API to
be stable!